// src/api/models/dto/NotificationResponse.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/

import type { DtoActorResponse } from "./ActorResponse.ts";

export type DtoNotificationResponse = {
    /**
     * @type object | undefined
    */
    actor?: DtoActorResponse;
    /**
     * @type string | undefined
    */
    created_at?: string;
    /**
     * @type array | undefined
    */
    group_actors?: DtoActorResponse[];
    /**
     * @type integer | undefined
    */
    group_count?: number;
    /**
     * @type string | undefined
    */
    id?: string;
    /**
     * @type boolean | undefined
    */
    is_read?: boolean;
    /**
     * @type string | undefined
    */
    message?: string;
    /**
     * @type string | undefined
    */
    secondary_id?: string;
    /**
     * @type string | undefined
    */
    target_id?: string;
    /**
     * @type string | undefined
    */
    type?: string;
};